userModels and procedure files refacored
